requirements.txt
modal==0.66.50
python-dotenv==1.0.1
slack_bolt==1.18.1